backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
faiss-cpu
PyPDF2
openai
pydantic
pickle-mixin
